chore: upgrade Rollup and related dependencies to latest version
