+++ package.json
@@ -31,7 +31,7 @@
   "scripts": {
     "prebuild": "yarn run clean",
     "build": "npm-run-all -s compile build:umd",
-    "build:umd": "rollup -c ./rollup.config.js",
+    "build:umd": "rollup -c ./rollup.config.mjs",
     "compile": "npm-run-all -p compile:es2015 compile:es5 compile:cjs compile:types",
     "compile:es2015": "tsc -p ./tsconfig.es2015.json",
     "compile:es5": "tsc -p ./tsconfig.es5.json",
@@ -48,17 +48,16 @@
     "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^22.0.0",
-    "@rollup/plugin-node-resolve": "^13.2.1",
+    "@rollup/plugin-commonjs": "^28.0.0",
+    "@rollup/plugin-node-resolve": "^15.3.0",
+    "@rollup/plugin-terser": "^0.4.4",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.30",
-    "camelcase": "^6.3.0",
+    "camelcase": "^8.0.0",
     "jest": "^28.0.3",
     "jest-environment-jsdom": "^28.0.2",
     "npm-run-all": "^4.1.5",
-    "rollup": "^2.70.2",
-    "rollup-plugin-sourcemaps": "^0.6.3",
-    "rollup-plugin-terser": "^7.0.2",
+    "rollup": "^4.24.0",
     "ts-jest": "next",
     "tslint": "^6.1.3",
     "typescript": "^4.6.4"
